// src/enums/errorMessages.ts
export enum ErrorMessages{
    DEFAULT_LOGIN = "Login failed. Please try again.",
    DEFAULT_SIGNUP = "Signup failed. Please try again.",
    NAME_REQUIRED = "Name is required",
    EMAIL_REQUIRED = "Email is required",
    PASSWORD_REQUIRED = "Password is required",
    NAME_MIN_LENGTH = "Name must be at least 3 characters",
    PASSWORD_MIN_LENGTH = "Password must be at least 8 characters",
    EMAIL_INVALID = "Email is invalid",
    PASSWORD_LETTER_REQUIRED = "Password must contain at least one letter",
    PASSWORD_NUMBER_REQUIRED = "Password must contain at least one number",
    PASSWORD_SPEC_CHAR_REQUIRED = "Password must contain at least one special character",

}